refactor: Move serialize_datetime function
Move serialize_datetime function to be in it's own file

Refactor #COG-170
